// cat-launcher/src/generated-types/ModsUpdatePayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GameVariant } from "./GameVariant";
import type { Mod } from "./Mod";
import type { ModsUpdateStatus } from "./ModsUpdateStatus";

/**
 * Payload sent when updating mod information.
 */
export type ModsUpdatePayload = { 
/**
 * The game variant the updates pertain to.
 */
variant: GameVariant, 
/**
 * The list of mods.
 */
mods: Array<Mod>, 
/**
 * The status of the update process.
 */
status: ModsUpdateStatus, };